Optimize the factory module pattern.
Use a named export for the main function and export the main function bound with it's dependencies (return of the factory) as default.

That way it is possible to easily import the ready to use function with it's dependencies bound for regular usage but also import the "raw" function without dependencies for testing.